Monster.txt Commit
Set up the format for monsters.txt, but I still need to make a script to
read it into some arrays and push those to the parent file. Also added
Testmonster to the monsters file, so mob_testmonster is now legacy and
is to be used for future reference only.