Actual Read Item Commit
This is the actual read item commit for the complete version. As I said
before, it needs some testing and probably re-writing, but I commented
it and I think it will work. It's based off the read script for
monsters.txt, which works very well, so it should work just fine.