movement system, abtract object, player object
This fork contains a different movement system, changes the player
object in relation to that, and some changes to the abstract objects. To
keep the player from jolting along the objects in the overworld, the
friction of those objects needs to be set to 0. This is because the
engine will otherwise cause objects to 'drag and stick' against each
other as they move to simulate friction.