Lined Up Words
There was a weird thing where the second and third lines were off by one
space, so I added one into the pp_dialog script right after the line
break. Flowey intro is done, it just needs to do a room transition into
the RESET or DO NOT room, which can be done later once saving and
loading is done.