More Item Menu Stuff
Fixing some minor stuff in the item menu. It displays stuff properly now
in the post-use menu screen, I just need to add in the sound for speech
and that's done. Still a bug when using items: draws Undefined for a
single frame after item use. Probably easy to fix, just re-order the
assignment of the item variable to be after the drawing.