Bug Fix
Fixed a bug where when you used an item it would draw "undefined" in
that item's place for a single frame. It just looked really weird, but
no problem now. I think that it should work perfectly now, unless we
want to do something different with items in a special fight.