Battle Animation
So, I realized that the way that I was doing battle animation was very
ham-fisted and was not going to work in more complicated battle
animations. I'm going to need to re-work it from the ground up later to
just move sprites back and forth along an x and y axis.